Move NametagCleanUpSystem after avatar clean up because of the dependency on DeferDeletion
